+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom costing logic for Project DocType to calculate `custom_sales_stock_cost` and `custom_incoming_material_value` from stock ledger entries, and dynamically update `expense_amount`.
 
 ### Changed
+- Moved Stock Entry from Purchase section to Project section in Project DocType dashboard.
 ### Deprecated
 ### Removed
 ### Fixed